mig: wait for the encrypted device to appear, and relax luksClose requirement on cleanup (boo#1225882)
On some machines it might take a bit for the mapped device to appear, so
properly wait for it (for 5 seconds maximum).

Do the same when cleaning up previously opened encrypted devices. Also,
do not ever fail if unable to close the device after 5 seconds (it's going
to be replaced anyway).

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>