Allow modifying fstab before mount
When /etc/fstab.repart exists, it is now used as fstab file.
After the system is installed, a special /etc/fstab.tik file can be used to inject or override fstab entries systemd-repart couldn't generate. It is expected that this file exists in the installed system right after deployment and before any post module runs.