Auto-detect system name
The name of the to be installed system is now automatically discovered, unless explicitely provided thrugh the config. For self deploy mode, /etc/os-release of the current system is read and ID and PRETTY_NAME are used. For raw images, Tik will try to read /etc/os-release to get these information or derive them from the image name after a 1s timeout/failure. The name from the images is only used if it is the same for all images available for tik.

This minimized aeon-specific code in tik-core and tik-functions