Refactor and fixes
Now that all functionality is implemented, it needed to be refactored to
make it more maintainable, easier to understand and easier to read.
In optional mode, not unlocking an encrypted partition no longer stops
the complete operation and all (encrypted) partition on a disk are now
checked for a root partition. probe_partitions also stops now after
finding the first match.